+++ Web/App/entities/elementField.js
@@ -26,6 +26,23 @@
             var _indexRating = null;
             var _currentUserIndexRating = null;
 
+            // Value filter for element cells
+            self.valueFilter = 1;
+            self.toggleValueFilter = function () {
+                self.valueFilter = self.valueFilter === 1 ? 2 : 1;
+
+                for (var i = 0; i < self.ElementCellSet.length; i++) {
+                    var cell = self.ElementCellSet[i];
+                    cell.setNumericValue();
+                    cell.setNumericValueMultiplied();
+                }
+
+                // $rootScope.$broadcast('elementValueFilterChanged', { element: self });
+            }
+            self.valueFilterText = function () {
+                return self.ElementCellSet[0].numericValueCount();
+            }
+
             // Other users' values: Keeps the values excluding current user's
             self.otherUsersIndexRating = null;
             self.otherUsersIndexRatingCount = null;